+++ src/handlers/runHandler.ts
@@ -1,4 +1,4 @@
-import { RunCommandOptions } from "../args";
+import { RunCommandOptions } from "../cliCommand";
 import { join, resolve, sep } from "path";
 import { EnantiomConfigLoader } from "../EnantiomConfigLoader";
 import { ScreenshotService } from "../ScreenshotService";
